Refactored; have rendering with chunks. Next: Face culling.

+++ native/src/rendering.rs
@@ -2,13 +2,13 @@ use euclid::default::*;
 
 #[derive(Clone, PartialEq, Debug)]
 pub struct Mesh {
-    pub index: Vec<u16>,
+    pub index: Vec<u32>,
     pub positions: Vec<Point3D<f32>>,
     pub normals: Vec<Vector3D<f32>>,
 }
 
 impl Mesh {
-    pub fn new(index: Vec<u16>, positions: Vec<Point3D<f32>>, normals: Vec<Vector3D<f32>>) -> Mesh {
+    pub fn new(index: Vec<u32>, positions: Vec<Point3D<f32>>, normals: Vec<Vector3D<f32>>) -> Mesh {
         Mesh {
             index,
             positions,
